perf:笔记本线上和本地抽公共组件和方法；feqat:搭建本地笔记本框架
