+++ app/renderer/src/main/src/pages/notepadManage/modifyNotepad/modifyNotepadLocal/ModifyNotepadLocal.tsx
@@ -1,0 +1,8 @@
+import React from "react"
+import {ModifyNotepadLocalProps} from "./ModifyNotepadLocalType"
+
+const ModifyNotepadLocal: React.FC<ModifyNotepadLocalProps> = React.memo((props) => {
+    return <div>编辑本地</div>
+})
+
+export default ModifyNotepadLocal
