+++ app/renderer/src/main/src/pages/notepadManage/modifyNotepad/modifyNotepadLocal/ModifyNotepadLocalType.d.ts
@@ -1,0 +1,1 @@
+export interface ModifyNotepadLocalProps {}
