+++ app/renderer/src/main/src/pages/notepadManage/modifyNotepad/modifyNotepadOnline/ModifyNotepadOnlineType.d.ts
@@ -1,0 +1,3 @@
+import {ModifyNotepadProps} from "../ModifyNotepadType"
+
+export interface ModifyNotepadOnlineProps extends ModifyNotepadProps {}
