+++ app/renderer/src/main/src/pages/notepadManage/notepadManage/notepadManageLocal/NotepadManageLocal.tsx
@@ -1,0 +1,8 @@
+import React from "react"
+import {NotepadManageLocalProps} from "./NotepadManageLocalType"
+
+const NotepadManageLocal: React.FC<NotepadManageLocalProps> = React.memo((props) => {
+    return <div>本地笔记本列表</div>
+})
+
+export default NotepadManageLocal
