+++ app/renderer/src/main/src/pages/notepadManage/notepadManage/notepadManageLocal/NotepadManageLocalType.d.ts
@@ -1,0 +1,1 @@
+export interface NotepadManageLocalProps {}
