+++ app/renderer/src/main/src/pages/notepadManage/notepadManage/notepadManageOnline/NotepadManageOnline.module.scss
@@ -1,0 +1,1 @@
+@use '../NotepadManage.module.scss'
