+++ app/renderer/src/main/src/pages/notepadManage/notepadManage/notepadManageOnline/NotepadManageOnlineType.d.ts
@@ -1,0 +1,3 @@
+import {NotepadManageProps} from "../NotepadManageType"
+
+export interface NotepadOnlineProps extends NotepadManageProps {}
